Migrating static website WIP
I am migrating the static Angular site hosted on S3 + CloudFront to CDK instead of CloudFormation. It appears that there will be some work in the direct translation due to OAC constructs not being available.

Once the migration is done, I'll refactor the CDK project since it will no longer be strictly backend.